cleans up a bunch of files and adds tests
