--- Common/Sources/Common/Extensions/UIColor+Hex.swift
@@ -1,42 +1,0 @@
-//
-//  UIColor+Hex.swift
-//  CocoaheadsConf
-//
-//  Created by Bruno Bilescky on 16/11/16.
-//  Copyright © 2016 Cocoaheads. All rights reserved.
-//
-
-import UIKit
-
-// TODO: Start using color assets instead of this
-
-extension UIColor {
-    
-    public convenience init(hexString: String) {
-        let hexString = hexString.trimmingCharacters(in: CharacterSet.whitespacesAndNewlines.union(CharacterSet(charactersIn: "#")))
-        let scanner   = Scanner(string: hexString)
-        var color: UInt32 = 0
-        
-        if scanner.scanHexInt32(&color) {
-            self.init(hex: color)
-        }
-        else {
-            self.init(hex: 0x000000)
-        }
-    }
-    
-    public convenience init(hex: UInt32) {
-        let mask = 0x000000FF
-        
-        let r = Int(hex >> 16) & mask
-        let g = Int(hex >> 8) & mask
-        let b = Int(hex) & mask
-        
-        let red   = CGFloat(r) / 255
-        let green = CGFloat(g) / 255
-        let blue  = CGFloat(b) / 255
-        
-        self.init(red:red, green:green, blue:blue, alpha:1)
-    }
-    
-}
