--- Common/Tests/CommonTests/CommonTests.swift
@@ -1,12 +1,0 @@
-import XCTest
-@testable import Common
-
-final class CommonTests: XCTestCase {
-    func testExample() throws {
-        // XCTest Documentation
-        // https://developer.apple.com/documentation/xctest
-
-        // Defining Test Cases and Test Methods
-        // https://developer.apple.com/documentation/xctest/defining_test_cases_and_test_methods
-    }
-}
