--- NSBrazilConf/Storage/Store.swift
@@ -1,6 +1,0 @@
-
-import Foundation
-
-public protocol Store {
-    func fetchInfo()
-}
